Wire AccountMenu into TopNav layout and update tests
