+++ components/shared/layout/index.ts
@@ -7,3 +7,4 @@ export { default as DashboardLayout } from './DashboardLayout';
 export { SideNav } from './SideNav';
 export { Breadcrumbs } from './Breadcrumbs';
 export type { BreadcrumbItem, BreadcrumbsProps } from './Breadcrumbs'; 
+export { default as AccountMenu } from './AccountMenu';
